# docs/openapi-public.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/leads": {
      "post": {
        "operationId": "LeadsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Leads"
        ]
      },
      "get": {
        "operationId": "LeadsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rating",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "leadSource",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minScore",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/stats": {
      "get": {
        "operationId": "LeadsController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/{id}": {
      "get": {
        "operationId": "LeadsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Leads"
        ]
      },
      "patch": {
        "operationId": "LeadsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLeadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Leads"
        ]
      },
      "delete": {
        "operationId": "LeadsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/{id}/score": {
      "post": {
        "operationId": "LeadsController_scoreLead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/{id}/convert": {
      "post": {
        "operationId": "LeadsController_convertLead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertLeadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/bulk/update": {
      "post": {
        "operationId": "LeadsController_bulkUpdate",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/bulk/delete": {
      "post": {
        "operationId": "LeadsController_bulkDelete",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/bulk/assign": {
      "post": {
        "operationId": "LeadsController_bulkAssign",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
...
    "/api/esignature/requests": {
      "post": {
        "operationId": "ESignatureController_createRequest",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      },
      "get": {
        "operationId": "ESignatureController_listRequests",
        "parameters": [
          {
            "name": "quoteId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
    "/api/esignature/requests/by-quote/{quoteId}": {
      "get": {
        "operationId": "ESignatureController_getByQuote",
        "parameters": [
          {
            "name": "quoteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
    "/api/esignature/requests/{id}": {
      "get": {
        "operationId": "ESignatureController_getRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
    "/api/esignature/requests/{id}/send": {
      "post": {
        "operationId": "ESignatureController_sendRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
    "/api/esignature/requests/{id}/refresh": {
      "post": {
        "operationId": "ESignatureController_refreshStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
    "/api/esignature/requests/{id}/void": {
      "post": {
        "operationId": "ESignatureController_voidRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
    "/api/esignature/requests/{id}/resend": {
      "post": {
        "operationId": "ESignatureController_resendRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
    "/api/esignature/requests/{id}/download": {
      "get": {
        "operationId": "ESignatureController_downloadSignedDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
    "/api/esignature/requests/{requestId}/signing-url/{signerId}": {
      "get": {
        "operationId": "ESignatureController_getSigningUrl",
        "parameters": [
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "signerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
    "/api/esignature/webhook/{provider}": {
      "post": {
        "operationId": "ESignatureController_handleWebhook",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
    "/api/esignature/stats": {
      "get": {
        "operationId": "ESignatureController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
    "/api/esignature/providers": {
      "get": {
        "operationId": "ESignatureController_getProviders",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "E-Signature"
        ]
      }
    },
...
  },
  "info": {
    "title": "SalesOS API",
    "description": "SalesOS is an AI-powered Sales CRM and Revenue Intelligence platform. This API provides programmatic access to CRM data, sales coaching, CPQ, and more.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://www.salesos.org/api",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key authentication"
      }
    },
    "schemas": {
...
      "CreateLeadDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateLeadDto": {
        "type": "object",
        "properties": {}
      },
      "ConvertLeadDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}